Split of multi files, update deps and add factory function to create self instance
