+++ package.json
@@ -4,13 +4,18 @@
   "description": "Asynchronous start of functions in JS. Supports priority and interrupt execution every 16 milliseconds, to achieve 60fps.",
   "scripts": {
     "build": "rollup -c",
+    "size": "npm run build && size-limit",
     "prepublishOnly": "npm run build",
-    "test": "jest",
+    "test": "npm run test:unit && npm run size",
+    "test:unit": "jest",
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
     "cli": "npm run test:coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
   },
-  "files": [ "dist", "src" ],
+  "files": [
+    "dist",
+    "src"
+  ],
   "main": "dist/frameScheduling.js",
   "module": "dist/frameScheduling.esm.js",
   "es2015": "dist/frameScheduling.esm2015.js",
@@ -20,6 +25,12 @@
     "url": "git+https://github.com/Tom910/frame-scheduling.git"
   },
   "keywords": [],
+  "size-limit": [
+    {
+      "path": "dist/frameScheduling.js",
+      "limit": "930 B"
+    }
+  ],
   "author": "Andey Marchenko <tom910ru@gmail.com>",
   "license": "MIT",
   "bugs": {
@@ -27,15 +38,16 @@
   },
   "homepage": "https://github.com/Tom910/frame-scheduling#readme",
   "devDependencies": {
-    "@types/jest": "^21.1.4",
-    "@types/node": "^8.0.46",
-    "coveralls": "^3.0.0",
-    "jest": "^21.2.1",
-    "prettier": "1.7.4",
-    "rollup": "^0.64.1",
-    "rollup-plugin-typescript2": "^0.16.1",
-    "ts-jest": "^21.1.4",
-    "tslint": "^5.8.0",
-    "typescript": "^3.0.1"
+    "@types/jest": "^24.0.13",
+    "@types/node": "^12.0.4",
+    "coveralls": "^3.0.3",
+    "jest": "^24.8.0",
+    "prettier": "1.17.1",
+    "rollup": "^1.13.1",
+    "rollup-plugin-typescript2": "^0.21.1",
+    "size-limit": "^1.3.5",
+    "ts-jest": "^24.0.2",
+    "tslint": "^5.17.0",
+    "typescript": "^3.5.1"
   }
 }
