+++ tsconfig.json
@@ -11,6 +11,7 @@
     "noImplicitThis": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "esModuleInterop": true,
     "declaration": true,
     "outDir": "./dist",
     "lib": ["es2015", "dom"],
